requirements.txt
# 浏览器自动化（Web插件支持）
playwright==1.40.0

# 图片处理
Pillow==10.1.0

# HTTP请求
requests==2.31.0

# 定时任务
APScheduler==3.10.4

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 配置文件
pyyaml==6.0.1

# Web框架
fastapi==0.109.0
uvicorn==0.27.0

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6